refactor: cleanup .gitignore
- move some more scripts
