+++ .gitignore
@@ -19,30 +19,12 @@ pids
 *.pid.lock
 *.swp
 
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# nyc test coverage
-.nyc_output
-
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
-# Bower dependency directory (https://bower.io/)
-bower_components
-
-# node-waf configuration
-.lock-wscript
 
 # Compiled binary addons (https://nodejs.org/api/addons.html)
 build/Release
 
 # Dependency directories
 node_modules/
-jspm_packages/
 
 # TypeScript v1 declaration files
 typings/
@@ -59,16 +41,11 @@ typings/
 # Output of 'npm pack'
 *.tgz
 
-# Yarn Integrity file
-.yarn-integrity
 
 # dotenv environment variables file
 .env
 .envrc
 .direnv
-
-# next.js build output
-.next
 
 # conversations
 static/conversations.css
@@ -81,9 +58,6 @@ embedded.provisionprofile
 
 stub.js
 
-# message list e2e test
-.test_tmp_data
-
 deltachat-rpc-server-aarch64-macos
 
 # nix files
@@ -95,3 +69,6 @@ packages/target-tauri/src-tauri/gen/schemas
 
 # macOS .DS_Store files
 .DS_Store
+
+# local-only dev scripts
+bin/_local/
