--- bin/lib/readme
@@ -1,1 +1,0 @@
-all files that are not executable but needed by the executable scripts in the bin/ folder
