+++ bin/link_core/link_version.js
@@ -30,7 +30,6 @@ const updatedContent = fileContent
 writeFileSync(filePath, updatedContent, 'utf8')
 
 execSync('pnpm i', { stdio: 'inherit' })
-execSync('./bin/link_core/link_catalog.sh', { stdio: 'inherit' })
 
 // cargo / tauri
 try {
