+++ bin/tools/copy.js
