+++ packages/frontend/package.json
@@ -9,12 +9,12 @@
     "build:ts": "node ./bin/build-frontend-ts.mjs",
     "build:source-map-support": "node ./bin/build-frontend-source_map_support.js",
     "build:scss": "sass ./scss/manifest.scss ./html-dist/main.css --embed-source-map --style=compressed",
-    "build:static": "node ../../bin/lib/copy.js ../../static ./html-dist && node ../../bin/lib/copy.js ./static ./html-dist && node ../../bin/lib/copy.js ../../images ./html-dist/images",
+    "build:static": "node ../../bin/tools/copy.js ../../static ./html-dist && node ../../bin/tools/copy.js ./static ./html-dist && node ../../bin/tools/copy.js ../../images ./html-dist/images",
     "build:themes": "sass themes/light.scss static/fallback-theme.css --style=compressed --embed-source-map && node ./bin/build-frontend-themes.js",
     "watch": "pnpm build:source-map-support && pnpm watch:ts & pnpm watch:static & pnpm watch:scss & pnpm watch:types",
     "watch:ts": "pnpm build:ts -- -w",
     "watch:types": "npx tsc --noEmit --pretty -w --preserveWatchOutput",
-    "watch:static": "node ../../bin/lib/copy.js ../../static ./html-dist & node ../../bin/lib/copy.js ./static ./html-dist & node ../../bin/lib/copy.js ../../images ./html-dist/images",
+    "watch:static": "node ../../bin/tools/copy.js ../../static ./html-dist & node ../../bin/tools/copy.js ./static ./html-dist & node ../../bin/tools/copy.js ../../images ./html-dist/images",
     "watch:scss": "pnpm build:scss --watch",
     "test": "NODE_OPTIONS=\"--enable-source-maps --import ./bin/register-ts-node.mjs\" mocha 'src/tests/**/*.ts'"
   },
