+++ packages/target-browser/package.json
@@ -10,7 +10,7 @@
     "build:locales": "pnpm -w translations:convert",
     "build:backend": "node ./bin/build.js",
     "build:runtime-impl": "pnpm esbuild --format=esm --bundle --minify --keep-names --sourcemap --outdir=./dist runtime-browser/runtime.ts",
-    "build:compose-frontend": "node ../../bin/lib/copy.js ../frontend/html-dist ./dist && node ../../bin/lib/copy.js ./static ./dist && pnpm build:runtime-impl",
+    "build:compose-frontend": "node ../../bin/tools/copy.js ../frontend/html-dist ./dist && node ../../bin/tools/copy.js ./static ./dist && pnpm build:runtime-impl",
     "start": "pnpm run build && node dist/server.js",
     "start:server": "node dist/server.js"
   },
