+++ packages/target-electron/package.json
@@ -34,11 +34,11 @@
     "build": "pnpm build:locales && pnpm build:backend && pnpm --filter=@deltachat-desktop/frontend build && pnpm build:compose-frontend && pnpm build:calls-webapp",
     "build:locales": "pnpm -w translations:convert",
     "build:backend": "node ./bin/build.js",
-    "build:calls-webapp": "node ../../bin/lib/copy.js ./node_modules/@deltachat/calls-webapp/dist ./html-dist/calls-webapp",
+    "build:calls-webapp": "node ../../bin/tools/copy.js ./node_modules/@deltachat/calls-webapp/dist ./html-dist/calls-webapp",
     "build:runtime-impl": "pnpm esbuild --format=esm --bundle --minify --keep-names --sourcemap --outdir=./html-dist runtime-electron/runtime.ts",
     "build:preload": "pnpm esbuild --platform=node --format=cjs --bundle --minify --keep-names --sourcemap --external:electron --outdir=./html-dist runtime-electron/preload.ts",
-    "build:compose-frontend": "node ../../bin/lib/copy.js ../frontend/html-dist ./html-dist && node ../../bin/lib/copy.js ./static ./html-dist && pnpm build:runtime-impl && pnpm build:preload",
-    "watch:compose-frontend": "node ../../bin/lib/copy.js ../frontend/html-dist ./html-dist -w & node ../../bin/lib/copy.js ./static ./html-dist -w & pnpm build:runtime-impl --watch & pnpm build:preload --watch",
+    "build:compose-frontend": "node ../../bin/tools/copy.js ../frontend/html-dist ./html-dist && node ../../bin/tools/copy.js ./static ./html-dist && pnpm build:runtime-impl && pnpm build:preload",
+    "watch:compose-frontend": "node ../../bin/tools/copy.js ../frontend/html-dist ./html-dist -w & node ../../bin/tools/copy.js ./static ./html-dist -w & pnpm build:runtime-impl --watch & pnpm build:preload --watch",
     "watch": "pnpm -w --filter=@deltachat-desktop/frontend watch & pnpm watch:compose-frontend",
     "dev": "pnpm build && pnpm start",
     "dev-inspect": "pnpm build && pnpm start --inspect",
