+++ packages/target-tauri/package.json
@@ -14,7 +14,7 @@
     "build:runtime-impl": "pnpm esbuild --format=esm --bundle --minify --keep-names --sourcemap --outdir=./html-dist runtime-tauri/runtime.ts",
     "build:html-email-view": "pnpm esbuild --bundle --minify --keep-names --sourcemap --outdir=./html-dist/tauri_html_email_view tauri-html-email-view/html_email_view.ts",
     "build:webxdc-js": "pnpm esbuild --bundle --minify --keep-names --outdir=./html-dist/webxdc webxdc-js-implementation/webxdc.ts",
-    "build:compose-frontend": "node ../../bin/lib/copy.js ../frontend/html-dist ./html-dist && node ../../bin/lib/copy.js ./static ./html-dist && pnpm build:runtime-impl",
+    "build:compose-frontend": "node ../../bin/tools/copy.js ../frontend/html-dist ./html-dist && node ../../bin/tools/copy.js ./static ./html-dist && pnpm build:runtime-impl",
     "tauri": "tauri",
     "fmt": "cd src-tauri && cargo fmt",
     "lint": "cd src-tauri && cargo clippy"
